ats_buddy_be/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.5
python-multipart==0.0.20
pypdf==5.2.0
python-docx==1.1.2
